+++ tests/unit/test_helper_subarray_leaf_device.py
@@ -31,6 +31,7 @@ commands_without_argin = [
 def test_leaf_node_command_with_argument(tango_context, command):
     dev_factory = DevFactory()
     subarray_leaf_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     result, command_id = subarray_leaf_device.command_inout(command, "")
     assert result[0] == ResultCode.QUEUED
     assert isinstance(command_id[0], str)
@@ -39,6 +40,7 @@ def test_leaf_node_command_with_argument(tango_context, command):
 def test_set_obs_state_transition(tango_context):
     dev_factory = DevFactory()
     subarray_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_device.adminMode = AdminMode.ONLINE
     subarray_device.AddTransition('[["CONFIGURING", 0.1]]')
     assert (
         subarray_device.obsStateTransitionDuration == '[["CONFIGURING", 0.1]]'
@@ -48,6 +50,7 @@ def test_set_obs_state_transition(tango_context):
 def test_reset_obs_state_transition(tango_context):
     dev_factory = DevFactory()
     subarray_leaf_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     subarray_leaf_device.AddTransition('[["CONFIGURING", 0.1]]')
     assert (
         subarray_leaf_device.obsStateTransitionDuration
@@ -60,6 +63,7 @@ def test_reset_obs_state_transition(tango_context):
 def test_obs_state_tranisition_for_configure(tango_context):
     dev_factory = DevFactory()
     subarray_leaf_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     subarray_leaf_device.AddTransition('[["READY", 0.0]]')
     _, _ = subarray_leaf_device.Configure("")
     wait_for_obstate(subarray_leaf_device, ObsState.READY)
@@ -69,6 +73,7 @@ def test_obs_state_tranisition_for_configure(tango_context):
 def test_obs_state_tranisition_for_assignresources(tango_context):
     dev_factory = DevFactory()
     subarray_leaf_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     subarray_leaf_device.AddTransition('[["READY", 0.0]]')
     _, _ = subarray_leaf_device.AssignResources("")
     wait_for_obstate(subarray_leaf_device, ObsState.READY)
@@ -78,6 +83,7 @@ def test_obs_state_tranisition_for_assignresources(tango_context):
 def test_clear_commandCallInfo(tango_context):
     dev_factory = DevFactory()
     subarray_leaf_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     _, _ = subarray_leaf_device.command_inout("Configure", "")
     subarray_leaf_device.command_inout("ClearCommandCallInfo")
     command_call_info = subarray_leaf_device.commandCallInfo
@@ -142,6 +148,7 @@ def test_assign_resources_command_not_allowed_after_queuing(
         "enabled": True,
         "fault_type": FaultType.COMMAND_NOT_ALLOWED_AFTER_QUEUING,
     }
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     subarray_leaf_device.SetDefective(json.dumps(defect))
     subarray_leaf_device.subscribe_event(
         "longRunningCommandResult",
@@ -166,6 +173,7 @@ def test_assign_resources_command_not_allowed_exception_after_queuing(
 ):
     dev_factory = DevFactory()
     subarray_leaf_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     defect = {
         "enabled": True,
         "fault_type": FaultType.COMMAND_NOT_ALLOWED_EXCEPTION_AFTER_QUEUING,
@@ -199,6 +207,7 @@ def test_assign_resources_command_not_allowed_exception_after_queuing(
 def test_assign_resources_stuck_in_intermediate_state(tango_context):
     dev_factory = DevFactory()
     subarray_leaf_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     defect = {
         "enabled": True,
         "fault_type": FaultType.STUCK_IN_INTERMEDIATE_STATE,
@@ -217,6 +226,7 @@ def test_assign_resources_stuck_in_intermediate_state(tango_context):
 def test_assign_resources_command_not_allowed(tango_context):
     dev_factory = DevFactory()
     subarray_leaf_device = dev_factory.get_device(SDP_LEAF_NODE_DEVICE)
+    subarray_leaf_device.adminMode = AdminMode.ONLINE
     defect = {
         "enabled": True,
         "fault_type": FaultType.COMMAND_NOT_ALLOWED_BEFORE_QUEUING,